refactor: extract avatar color and initials utilities
Create src/lib/avatar.ts with shared utilities:
- avatarColor(userId): hash-based color for auth pages (8 colors)
- guestAvatarColor(index): index-based color for guest pages (8 colors)
- getInitials(name, email): initials from name or email fallback

Remove duplicated code from 5 files:
- dashboard/page.tsx: avatarColors, avatarColor, getInitials
- groups/[groupId]/page.tsx: avatarColors, avatarColor, getInitials
- split/page.tsx: inline initials function
- split/[token]/page.tsx: inline initials + unused colors array
- split/[token]/claim/page.tsx: colors array + initials function
