+++ src/app/[locale]/(app)/groups/[groupId]/page.tsx
@@ -27,37 +27,7 @@ import { Link } from "@/i18n/navigation";
 import { toast } from "sonner";
 import { InviteDialog } from "@/components/groups/invite-dialog";
 import { SettleDialog } from "@/components/groups/settle-dialog";
-
-function getInitials(name?: string | null, email?: string | null): string {
-  if (name) {
-    return name
-      .split(" ")
-      .map((n) => n[0])
-      .join("")
-      .toUpperCase()
-      .slice(0, 2);
-  }
-  return email?.[0]?.toUpperCase() ?? "?";
-}
-
-const avatarColors = [
-  "bg-blue-500",
-  "bg-emerald-500",
-  "bg-violet-500",
-  "bg-amber-500",
-  "bg-rose-500",
-  "bg-cyan-500",
-  "bg-fuchsia-500",
-  "bg-lime-500",
-];
-
-function avatarColor(userId: string): string {
-  let hash = 0;
-  for (let i = 0; i < userId.length; i++) {
-    hash = (hash * 31 + userId.charCodeAt(i)) | 0;
-  }
-  return avatarColors[Math.abs(hash) % avatarColors.length];
-}
+import { getInitials, avatarColor } from "@/lib/avatar";
 
 export default function GroupDetailPage({
   params,
